Refactor PartitionCompletenessValidator
